Feat/remaining extensions

+++ src/MES.AI.Application/Services/EquipmentHealthService.cs
@@ -106,8 +106,8 @@ public class EquipmentHealthService : IEquipmentHealthService
         foreach (var group in grouped)
         {
             var dayReports = group.ToList();
-            var goodQty = dayReports.Sum(r => r.GoodQty.Value);
-            var badQty = dayReports.Sum(r => (r.ScrapQty + r.ReworkQty).Value);
+            var goodQty = (double)dayReports.Sum(r => r.GoodQty.Value);
+            var badQty = (double)dayReports.Sum(r => (r.ScrapQty + r.ReworkQty).Value);
             var actualRunMinutes = dayReports.Sum(r => r.DurationMin);
 
             double plannedRunMinutes = equipment?.PlannedRunTime.HasValue == true && equipment.PlannedRunTime.Value > 0
